# openapi.spec.json
{
  "components": {
    "schemas": {
      "Todo": {
        "properties": {
          "content": {
            "minLength": 1,
            "maxLength": 1024,
            "type": "string"
          },
          "id": {
            "minLength": 1,
            "maxLength": 50,
            "type": "string"
          }
        },
        "type": "object",
        "required": [
          "content",
          "id"
        ]
      }
    }
  },
  "info": {
    "title": "",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/ping": {
      "get": {
        "operationId": "RootController.ping",
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Ping",
        "tags": [
          "Root"
        ]
      }
    },
    "/api/tag": {
      "get": {
        "operationId": "RootController.tag",
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Tag",
        "tags": [
          "Root"
        ]
      }
    },
    "/api/todo/{id}": {
      "get": {
        "operationId": "TodoController.getOne",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Get one",
        "tags": [
          "Todo"
        ]
      }
    },
    "/api/todo/": {
      "get": {
        "operationId": "TodoController.getAll",
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Get all",
        "tags": [
          "Todo"
        ]
      },
      "post": {
        "operationId": "TodoController.post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Todo"
              }
            }
          },
          "description": "Todo",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Post",
        "tags": [
          "Todo"
        ]
      }
    }
  }
}